Introduced CharsConfigurator
