+++ src/main/java/net/jqwik/providers/CharacterArbitraryProvider.java
@@ -5,7 +5,6 @@ import java.util.function.*;
 
 import net.jqwik.api.*;
 import net.jqwik.api.arbitraries.*;
-import net.jqwik.api.constraints.*;
 import net.jqwik.api.providers.*;
 
 public class CharacterArbitraryProvider implements ArbitraryProvider {
@@ -17,17 +16,6 @@ public class CharacterArbitraryProvider implements ArbitraryProvider {
 	@Override
 	public CharacterArbitrary provideFor(GenericType targetType, Function<GenericType, Optional<Arbitrary<?>>> subtypeProvider) {
 		return Arbitraries.chars();
-	}
-
-	public CharacterArbitrary configure(CharacterArbitrary arbitrary, Chars chars) {
-		return arbitrary.withChars(chars.value()).withChars(chars.from(), chars.to());
-	}
-
-	public CharacterArbitrary configure(CharacterArbitrary arbitrary, CharsList charsList) {
-		for (Chars chars : charsList.value()) {
-			arbitrary = configure(arbitrary, chars);
-		}
-		return arbitrary;
 	}
 
 }
