+++ src/main/resources/META-INF/services/net.jqwik.api.providers.ArbitraryConfigurator
@@ -1,3 +1,4 @@
 net.jqwik.configurators.WithNullConfigurator
 net.jqwik.configurators.ScaleConfigurator
 net.jqwik.configurators.SizeConfigurator
+net.jqwik.configurators.CharsConfigurator
